+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 1.0.1 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Register standard blocks serializers/deserializer for IDexterityContent. Otherwise they don't work.
+  [cekk]
 
 
 1.0.0 (2021-02-24)